ffmpeg_video_cover.sh calculate md5s

+++ .local/scripts/bin/ffmpeg_video_cover.sh
@@ -85,6 +85,8 @@ cleanup() {
     printf "Restoring backup... "
     if mv "$video_backup" "$video"; then
       echo "OK."
+    else
+      return
     fi
   else
     printf "Removing backup... "
@@ -104,7 +106,29 @@ cleanup() {
 }
 trap cleanup EXIT
 
+original_hash=
+new_hash=
+
 mv "$video" "$video_backup" || exit
+if ! original_hash="$(ffmpeg -i "$video_backup" -map 0:V -c copy -f md5 - 2>/dev/null)"; then
+  echo "Could not calculate original md5 hash" >&2
+  exit 1
+fi
+
 ffmpeg -y -i "$video_backup" -i "$cover" -map 1 -map 0:V -c copy -disposition:0 attached_pic "$out" || exit
 cp "$out" "$video" || exit
+if ! new_hash="$(ffmpeg -i "$out" -map 0:V -c copy -f md5 - 2>/dev/null)"; then
+  echo "Could not calculate new md5 hash" >&2
+  exit 1
+fi
+
+echo "Original video hash: $original_hash"
+echo "New video hash: $new_hash"
+if [[ "$original_hash" != "$new_hash" ]]; then
+  echo "Hashes don't match, aborting." >&2
+  exit 1
+else
+  echo "Hashes match!"
+fi
+
 done=1
